# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Collaborative Task Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  manager: "manager",
  user: "user",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type SignupRequestRole =
  (typeof SignupRequestRole)[keyof typeof SignupRequestRole];

export const SignupRequestRole = {
  manager: "manager",
  user: "user",
} as const;

export interface SignupRequest {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role?: SignupRequestRole;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority];

export const TaskPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Task {
  id: number;
  title: string;
  description?: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  dueDate?: string | null;
  createdAt: string;
  updatedAt: string;
  createdBy: User;
  assignedTo?: User | null;
}

export type CreateTaskRequestPriority =
  (typeof CreateTaskRequestPriority)[keyof typeof CreateTaskRequestPriority];

export const CreateTaskRequestPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface CreateTaskRequest {
  title: string;
  description?: string | null;
  priority?: CreateTaskRequestPriority;
  assignedToId?: number | null;
  dueDate?: string | null;
}

export type UpdateTaskRequestStatus =
  (typeof UpdateTaskRequestStatus)[keyof typeof UpdateTaskRequestStatus];

export const UpdateTaskRequestStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export type UpdateTaskRequestPriority =
  (typeof UpdateTaskRequestPriority)[keyof typeof UpdateTaskRequestPriority];

export const UpdateTaskRequestPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface UpdateTaskRequest {
  title?: string;
  description?: string | null;
  status?: UpdateTaskRequestStatus;
  priority?: UpdateTaskRequestPriority;
  assignedToId?: number | null;
  dueDate?: string | null;
}

export interface TaskListResponse {
  tasks: Task[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export type ActivityLogTask = {
  id: number;
  title: string;
};

export interface ActivityLog {
  id: number;
  taskId: number;
  userId: number;
  action: string;
  details?: string | null;
  createdAt: string;
  user: User;
  task: ActivityLogTask;
}

export interface ActivityLogListResponse {
  logs: ActivityLog[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export type GetTasksParams = {
  page?: number;
  limit?: number;
  status?: GetTasksStatus;
  assignedToMe?: boolean;
  createdByMe?: boolean;
};

export type GetTasksStatus =
  (typeof GetTasksStatus)[keyof typeof GetTasksStatus];

export const GetTasksStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export type GetActivityLogsParams = {
  taskId?: number;
  page?: number;
  limit?: number;
};