# lib/api-zod/src/generated/types/activityLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Collaborative Task Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityLogTask } from "./activityLogTask";
import type { User } from "./user";

export interface ActivityLog {
  id: number;
  taskId: number;
  userId: number;
  action: string;
  details?: string | null;
  createdAt: Date;
  user: User;
  task: ActivityLogTask;
}