# lib/api-zod/src/generated/types/activityLogListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Collaborative Task Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityLog } from "./activityLog";

export interface ActivityLogListResponse {
  logs: ActivityLog[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}